+++ app/src/main/java/dev/hossain/mathtutor/ui/onboarding/OnboardingScreen.kt
@@ -7,11 +7,14 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.asPaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBars
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
@@ -166,6 +169,7 @@ fun OnboardingContent(
     val pagerState = rememberPagerState(pageCount = { onboardingPages.size })
     val coroutineScope = rememberCoroutineScope()
     val currentPage = onboardingPages[pagerState.currentPage]
+    val systemBarsPadding = WindowInsets.systemBars.asPaddingValues()
 
     Box(
         modifier =
@@ -177,6 +181,7 @@ fun OnboardingContent(
             modifier =
                 Modifier
                     .fillMaxSize()
+                    .padding(systemBarsPadding)
                     .padding(24.dp),
             horizontalAlignment = Alignment.CenterHorizontally,
             verticalArrangement = Arrangement.SpaceBetween,